Moved enhanceWorkflows func to workflows.go
