+++ deployments/workflows.go
@@ -23,6 +23,7 @@ import (
 	"github.com/hashicorp/consul/api"
 	"github.com/pkg/errors"
 
+	"github.com/ystia/yorc/v4/deployments/internal"
 	"github.com/ystia/yorc/v4/helper/consulutil"
 	"github.com/ystia/yorc/v4/tosca"
 )
@@ -182,4 +183,49 @@ func DeleteWorkflow(kv *api.KV, deploymentID, workflowName string) error {
 	_, err := kv.DeleteTree(path.Join(consulutil.DeploymentKVPrefix, deploymentID,
 		workflowsPrefix, workflowName), nil)
 	return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
+}
+
+func enhanceWorkflows(consulStore consulutil.ConsulStore, kv *api.KV, deploymentID string) error {
+	wf, err := ReadWorkflow(kv, deploymentID, "run")
+	if err != nil {
+		return err
+	}
+	var wasUpdated bool
+	for sn, s := range wf.Steps {
+		var isCancellable bool
+		for _, a := range s.Activities {
+			switch strings.ToLower(a.CallOperation) {
+			case tosca.RunnableSubmitOperationName, tosca.RunnableRunOperationName:
+				isCancellable = true
+			}
+		}
+		if isCancellable && len(s.OnCancel) == 0 {
+			// Cancellable and on-cancel not defined
+			// Check if there is an cancel op
+			hasCancelOp, err := IsOperationImplemented(kv, deploymentID, s.Target, tosca.RunnableCancelOperationName)
+			if err != nil {
+				return err
+			}
+			if hasCancelOp {
+				cancelStep := &tosca.Step{
+					Target:             s.Target,
+					TargetRelationShip: s.TargetRelationShip,
+					OperationHost:      s.OperationHost,
+					Activities: []tosca.Activity{
+						tosca.Activity{
+							CallOperation: tosca.RunnableCancelOperationName,
+						},
+					},
+				}
+				csName := "yorc_automatic_cancellation_of_" + sn
+				wf.Steps[csName] = cancelStep
+				s.OnCancel = []string{csName}
+				wasUpdated = true
+			}
+		}
+	}
+	if wasUpdated {
+		internal.StoreWorkflow(consulStore, deploymentID, "run", wf)
+	}
+	return nil
 }
